% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degroot.R
\name{degroot}
\alias{degroot}
\title{Calculates the simple weighted avergaging model}
\usage{
degroot(W, Y, iterations, all.iter = F, row.normalize = T,
  self.weight = 1)
}
\arguments{
\item{W:}{the weight matrix to use}

\item{Y:}{the starting positions to use}

\item{iterations:}{the number of iterations to take}

\item{all.iter:}{T/F of whether all the iterations are returned in a matrix}

\item{row.normalize:}{T/F indicator of whether the matrix should be row
normalized before using}

\item{self.weight:}{numeric value(s) in [0, 1], indicating how much each
person weights him-/herself vs. others.  0 indicates full self-weight, and 
1 indicates full other-weight.  Only used if row.normalize is TRUE.}
}
\value{
A vector with the weighted average taken (iter) times
}
\description{
Calculates the simple weighted avergaging model
}
\details{
Uses the DeGroot weighted averaging model, which just returns
(W ^ iterations) %*% Y.  If row.normalize is true, set the diagonal equal to
1 and divide each row by it's sum.  Otherwise, throw a warning if each row
doesn't sum to 1.
}
\note{
Do NOT use DTMC for this -- DTMC requires that the Y variable sum to 1.
}
