% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degroot_list.R
\name{degrootList}
\alias{degrootList}
\title{Run the degroot function on a series transition matrices}
\usage{
degrootList(mat.list, Y, all.iter = T, row.normalize = T, self.weight = 1)
}
\arguments{
\item{mat.list:}{a list of matrices}

\item{Y:}{the initial values of the diffusion process}

\item{all.iter:}{T/F indicator of whether every iteration of the diffusion
process should be saved}

\item{row.normalize:}{T/F, should the matrices be row normalized before
proceeding?}

\item{self.weight:}{numeric value(s) in [0, 1], indicating how much each
person weights him-/herself vs. others.  0 indicates full self-weight, and 
1 indicates full other-weight.  Only used if row.normalize is TRUE.}
}
\value{
If all.iter is TRUE, a matrix of Y X length(mat.list), with the 
outcome of the diffusion process at each iteration.  If all.iter is FALSE,
the outcome of the last iteration
}
\description{
Run the degroot function on a series transition matrices
}
