% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_construct_matrix.R
\name{makeConstructMatrix}
\alias{makeConstructMatrix}
\title{Row-normalizes a matrix with arbitrary weighting for the diagonal}
\usage{
makeConstructMatrix(net, a)
}
\arguments{
\item{net:}{a statnet network object or a square matrix}

\item{a:}{a numeric value indicating how much weight to give to self vs.
other.  1 indicates full self-weight (diagonal), 0 indicates full other-
weight (off diagonal).  Can be a vector to give each person different
weights.}
}
\value{
a square matrix whose rows sum to 1 or 0
}
\description{
Row-normalizes a matrix with arbitrary weighting for the diagonal
}
\note{
Calculates the construct matrix from Friedkin and Johnson (2011:40).
That matrix is defined as \eqn{AC + I - A}, where:
\deqn{C_{ij} = \frac{W_{ij}}{\sum_{k: i \ne k} W_{ik}}}
\eqn{C_{ii} = 0}, \eqn{A_{ii} = a_i}, and \eqn{A_{ij} = 0} for all
\eqn{i \ne j}.

Among indegree isolates, \eqn{\sum_{k: i \ne k} W_{ik} = 0}.  In those cases,
\eqn{C_{ij} \coloneqq 0}, and \eqn{a_{i} \coloneqq 1}.
}
\references{
Friedkin, Noah E. and Eugene c. Johnson. 2011. 
\emph{Social Influence Network Theory: A Sociological Examination of Small
Group Dynamics}. Cambridge University Press: New York.
}
