% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_degroot_list.R
\name{tidyDegrootList}
\alias{tidyDegrootList}
\title{Wrapper function for tidyDegroot that calls tidyDegroot on a list of degroot
matrices}
\usage{
tidyDegrootList(pred, id)
}
\arguments{
\item{pred:}{a list of matrices, generally produced by ergmmDegroot}

\item{id:}{a vector of ID values that will be added as a column to the
resulting data.frame}
}
\value{
a data.frame (in tibble form) with 4 columns: the 3 columns produced
by tidyDegroot, plus an additional column for the number of draws, which 
denotes the index from the original list
}
\description{
Wrapper function for tidyDegroot that calls tidyDegroot on a list of degroot
matrices
}
